Refactored code for getting weekly booking requests
