+++ src/routes/(app)/booking/admin/+page.server.ts
@@ -1,25 +1,14 @@
 import apiNames from "$lib/utils/apiNames";
 import { authorize } from "$lib/utils/authorization";
-import dayjs from "dayjs";
 import type { PageServerLoad } from "./$types";
-import { actions } from "./sharedActions";
+import { actions, getAllBookingRequestsWeekly } from "./sharedUtils";
 
 export const load: PageServerLoad = async ({ locals }) => {
   const { prisma, user } = locals;
   authorize(apiNames.BOOKINGS.UPDATE, user);
 
-  const bookingRequests = await prisma.bookingRequest.findMany({
-    where: {
-      start: {
-        gte: dayjs().subtract(1, "week").toDate(),
-      },
-    },
-    orderBy: [{ start: "asc" }, { end: "asc" }, { status: "asc" }],
-    include: {
-      bookables: true,
-      booker: true,
-    },
-  });
+  const bookingRequests = await getAllBookingRequestsWeekly(prisma)
+
   return { bookingRequests };
 };
